% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_sample_occurrence.R
\name{cross_sample_occurrence}
\alias{cross_sample_occurrence}
\title{Sample cross-sample occurrence}
\usage{
cross_sample_occurrence(
  df,
  output.type = c("by_gene", "summary"),
  outlier.type = c("UPPER", "LOWER", "NONE")
)
}
\arguments{
\item{df}{The dataframe generated with \code{random_frag_model()} or \code{pattern_frag_model()} that contains the outlier information.}

\item{output.type}{Determines whether the output is the cross-sample occurrence across each gene in each sample (\code{"by_gene"}) or a summary of the cross-sample occurrences across samples (\code{"summary"}).}

\item{outlier.type}{Determines whether cross-sample occurrence is calculated for each outlier type (\code{"UPPER"}, \code{"LOWER"}, \code{"NONE"}).

Note that in order to obtain unique non-outliers, the \code{"NONE"} output will have to be further filtered to remove any/all overlap with the \code{"UPPER"} and \code{"LOWER"} outputs.}
}
\description{
Calculates the overlap in outlier genes across each sample.
}
\examples{
cross_sample_occurrence()

}
