% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cso_plot.R
\name{cso_plot}
\alias{cso_plot}
\title{Visualization of cross-sample occurrence}
\usage{
cso_plot(
  df,
  sample.num,
  sample.names,
  gene.type = c("RIGs", "RAGs"),
  write.out = c("yes", "no"),
  h = NULL,
  w = NULL,
  output.dir = NULL,
  file.name = NULL
)
}
\arguments{
\item{df}{The dataframe generated with \code{random_frag_model()}}

\item{sample.num}{The number of samples being analyzed. This is used to extend the color palette in case there are more samples than colors.}

\item{sample.names}{Character string defining the sample names along the x-axis.}

\item{gene.type}{Used for the y-axis title, denotes either \code{"RIGs"} or \code{"RAGs"}.}

\item{write.out}{Determines whether or not to save the generated plot. To save the plots, define as \code{"yes"}.}

\item{h}{The desired height of the generated plot}

\item{w}{The desired width of the generated plot}

\item{output.dir}{The directory in which the files will be output. If the working directory is the desired directory, enter \code{NULL}.}

\item{file.name}{The name of the output file. Enter as a character string.}
}
\description{
Plots the cross-sample overlap between samples.
}
\examples{
cso_plot()

}
