% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_file_list.R
\name{generate_file_list}
\alias{generate_file_list}
\title{File list construction}
\usage{
generate_file_list(directory, pattern)
}
\arguments{
\item{directory}{The directory to search. Entered as a character string. Can be called with \code{getwd()} if the directory of interest is the current working directory.
The function is written to search recursively, so all subdirectories within the named directory are searched.}

\item{pattern}{The common file naming pattern to search for. Input as a character string. Typically something like \code{"*genic.bed"}.}
}
\description{
Generates a list of integration site files in a given directory for import into rigrag.
}
\examples{
generate_file_list()

}
