% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genic_subpop.R
\name{genic_subpop}
\alias{genic_subpop}
\title{Quantifies the relative proportions of each outlier type per sample}
\usage{
genic_subpop(df, output.type = c("relative", "absolute"))
}
\arguments{
\item{df}{The dataframe generated with the model calculation functions}

\item{output.type}{Defines whether the output values are \code{"relative"} or \code{"absolute"}.
\code{"relative"} values express the population fractions relative to total integration.
\code{"absolute"} values expression the population fraction relative to genic integration.}
}
\description{
Quantifies the relative proportions of each outlier type per sample
}
\examples{
genic_subpop()

}
