% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genic_subpop_plot.R
\name{genic_subpop_plot}
\alias{genic_subpop_plot}
\title{Quantifies and plots the relative proportions of each outlier type per sample}
\usage{
genic_subpop_plot(
  df,
  output.type = c("relative", "absolute"),
  sample.levels,
  sample.names
)
}
\arguments{
\item{df}{The dataframe generated with the model calculation functions}

\item{output.type}{Defines whether the output values are \code{"relative"} or \code{"absolute"}.
\code{"relative"} values express the population fractions relative to total integration.
\code{"absolute"} values expression the population fraction relative to genic integration.}

\item{sample.levels}{Character string defining the factor levels of the samples. Used for ordering the plot.}

\item{sample.names}{Character string defining the sample names along the x-axis.}
}
\description{
Quantifies and plots the relative proportions of each outlier type per sample
}
\examples{
genic_subpop_plot()

}
