% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_file_import.R
\name{int_file_import}
\alias{int_file_import}
\title{Integration site file import and summary statistics}
\usage{
int_file_import(files, sample.ids, collapsed = c(TRUE, FALSE))
}
\arguments{
\item{files}{List of integration site files containing the data to be analyzed. Generated with \code{generate_file_list()}}

\item{sample.ids}{List of sample identifiers for each file in the file list.
Defined with \code{list()}.
The order of the sample identifiers must match the order of the files listed in the file list.
Spaces and special characters like '+' and '#' are OK in the sample identifiers.
Files written within package functions named using sample identifiers will have their names cleaned to remove spaces and select special characters.}

\item{collapsed}{If \code{"TRUE"}, the number of occurrences of each gene per sample are counted and the collapsed data is returned. If \code{"FALSE"}, every integration event into every gene is returned.}
}
\description{
Imports integration site files.
Calculates the number of genic integration sites and the fraction of genic integration events per sample.
Calculates gene length per gene.
}
\examples{
int_file_import()

}
