% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers_output.R
\name{outliers_output}
\alias{outliers_output}
\title{Single file outliers output}
\usage{
outliers_output(df, output.dir = NULL, file.prefix)
}
\arguments{
\item{df}{The dataframe generated with \code{random_frag_model()} that contains the outlier information.}

\item{output.dir}{The directory in which the file will be output. If the working directory is the desired directory, enter \code{NULL}.}

\item{file.prefix}{The desired prefix of the output files as a character string. The function will add "upper/lower_outliers.txt" to the prefix.}
}
\description{
Outputs a single unique gene list for the combined upper and lower outliers for all samples
}
\examples{
outliers_output()

}
