% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patient_overrep.R
\name{patient_overrep}
\alias{patient_overrep}
\title{Finds overrepresented genes in patient-derived integration datasets}
\usage{
patient_overrep(df, df2, column)
}
\arguments{
\item{column}{Name of the column containing the patient integration frequency values}

\item{df1}{Dataframe containing patient-derived genes and the corresponding integration frequency values. This should be a "final" list, with all averaging already done (if necessary).}
}
\description{
Compares patient integration frequencies to the corresponding frequencies in in vitro samples.
}
\examples{
patient_overrep()

}
