% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern_frag_model.R
\name{pattern_frag_model}
\alias{pattern_frag_model}
\title{Model generation based on pattern fragmentation}
\usage{
pattern_frag_model(df, pattern = c("NASB", "MB"))
}
\arguments{
\item{df}{The dataframe generated with \code{int_file_import}}

\item{pattern}{The character string defining the enzymes used to digest the genomic DNA. Currenty available patterns are \code{"MB"} for MseI/BglII digestion and \code{"NASB"} for NheI/AvrII/SpeI/BamHI digestion.}
}
\description{
Calculates the expected integration frequency, the upper boundary, and the lower boundary for each sample.
Defines outliers based on the relation to the calculated upper and lower boundaries.
}
\examples{
pattern_frag_model()

}
