% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_frag_model.R
\name{random_frag_model}
\alias{random_frag_model}
\title{Model generation based on random fragmentation}
\usage{
random_frag_model(df)
}
\arguments{
\item{df}{The dataframe generated with \code{int_file_import}}

\item{total.sites}{The column in the \code{df} dataframe containing the total number of integration sites.}

\item{gene.length}{The column in the \code{df} dataframe containing the gene length of a given gene.}
}
\description{
Calculates the expected integration frequency, the upper boundary, and the lower boundary for each sample.
Defines outliers based on the relation to the calculated upper and lower boundaries.
}
\examples{
random_frag_model()

}
