% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_outliers_output.R
\name{separate_outliers_output}
\alias{separate_outliers_output}
\title{Sample-specific outliers output}
\usage{
separate_outliers_output(
  df,
  outliers = c("both", "upper", "lower"),
  output.dir = NULL
)
}
\arguments{
\item{df}{The dataframe generated with \code{random_frag_model()} that contains the outlier information.}

\item{outliers}{The type of outliers that you want output (\code{"both"}, \code{"upper"}, or \code{"lower"}).}

\item{output.dir}{The directory in which the files will be output. If the working directory is the desired directory, enter \code{NULL}.}

\item{data.depth}{\code{"list"} returns a list of genes classified as outliers for each sample. \code{complete} returns all variables for each sample.}
}
\description{
Outputs a gene list for each sample analyzed and the specified outlier types.
}
\examples{
separate_outliers_output()

}
