% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total_cso_plot.R
\name{total_cso_plot}
\alias{total_cso_plot}
\title{Visualization of cross-sample occurrence}
\usage{
total_cso_plot(df, sample.num, population.names)
}
\arguments{
\item{df}{The dataframe generated with \code{random_frag_model()}}

\item{sample.num}{The number of samples being analyzed. This is used to extend the color palette in case there are more samples than colors.}

\item{population.names}{Character string defining the sample names along the x-axis.}
}
\description{
Plots the cross-sample overlap between samples.
}
\examples{
total_cso_plot()

}
