% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total_cso.R
\name{total_cso}
\alias{total_cso}
\title{Sample cross-sample occurrence}
\usage{
total_cso(df, uo_list, lo_list)
}
\arguments{
\item{df}{The dataframe generated with \code{random_frag_model()} or \code{pattern_frag_model()} that contains the outlier information.}

\item{uo_list}{The combined list of genes corresponding to upper outliers. Requires a "gene.name" column header. Easily generated using the \code{cross_sample_occurrence()} function.}

\item{lo_list}{The combined list of genes corresponding to lower outliers. Requires a "gene.name" column header. Easily generated using the \code{cross_sample_occurrence()} function.}
}
\description{
Calculates the overlap in outlier genes across each sample.
}
\examples{
total_cso()

}
